+++ package.json
@@ -60,8 +60,9 @@
     "unified": "6.1.6"
   },
   "devDependencies": {
-    "babel-cli": "6.24.1",
-    "babel-preset-es2015": "6.24.1",
+    "@babel/core": "7.0.0-beta.40",
+    "@babel/preset-es2015": "7.0.0-beta.40",
+    "builtin-modules": "2.0.0",
     "codecov": "2.2.0",
     "cross-env": "5.0.5",
     "eslint": "4.18.2",
@@ -82,11 +83,11 @@
     "rollup-plugin-node-globals": "1.1.0",
     "rollup-plugin-node-resolve": "2.0.0",
     "rollup-plugin-replace": "1.2.1",
+    "rollup-plugin-uglify": "3.0.0",
     "shelljs": "0.8.1",
     "snapshot-diff": "0.2.2",
     "strip-ansi": "4.0.0",
     "tempy": "0.2.1",
-    "uglify-es": "3.3.9",
     "webpack": "2.6.1"
   },
   "scripts": {
