+++ src/common/util.js
@@ -240,18 +240,6 @@ function hasSpaces(text, index, opts) {
   return idx !== index;
 }
 
-// Super inefficient, needs to be cached.
-function lineColumnToIndex(lineColumn, text) {
-  let index = 0;
-  for (let i = 0; i < lineColumn.line - 1; ++i) {
-    index = text.indexOf("\n", index) + 1;
-    if (index === -1) {
-      return -1;
-    }
-  }
-  return index + lineColumn.column;
-}
-
 function setLocStart(node, index) {
   if (node.range) {
     node.range[0] = index;
@@ -807,7 +795,6 @@ module.exports = {
   printNumber,
   hasIgnoreComment,
   hasNodeIgnoreComment,
-  lineColumnToIndex,
   makeString,
   addLeadingComment,
   addDanglingComment,
