+++ src/language-html/index.js
@@ -22,27 +22,11 @@ const languages = [
   }
 ];
 
-const parsers = {
-  parse5: {
-    get parse() {
-      return eval("require")("./parser-parse5");
-    },
-    astFormat: "htmlparser2",
-    locEnd: function(node) {
-      return node.__location && node.__location.endOffset;
-    },
-    locStart: function(node) {
-      return node.__location && node.__location.startOffset;
-    }
-  }
-};
-
 const printers = {
   htmlparser2: printer
 };
 
 module.exports = {
   languages,
-  parsers,
   printers
 };
