+++ src/language-js/index.js
@@ -2,67 +2,10 @@
 
 const estreePrinter = require("./printer-estree");
 const estreeJsonPrinter = require("./printer-estree-json");
-const hasPragma = require("./pragma").hasPragma;
 const options = require("./options");
-const privateUtil = require("../common/util");
 
 // Based on:
 // https://github.com/github/linguist/blob/master/lib/linguist/languages.yml
-
-const locStart = function(node) {
-  // Handle nodes with decorators. They should start at the first decorator
-  if (
-    node.declaration &&
-    node.declaration.decorators &&
-    node.declaration.decorators.length > 0
-  ) {
-    return locStart(node.declaration.decorators[0]);
-  }
-  if (node.decorators && node.decorators.length > 0) {
-    return locStart(node.decorators[0]);
-  }
-
-  if (node.__location) {
-    return node.__location.startOffset;
-  }
-  if (node.range) {
-    return node.range[0];
-  }
-  if (typeof node.start === "number") {
-    return node.start;
-  }
-  if (node.loc) {
-    return node.loc.start;
-  }
-  return null;
-};
-
-const locEnd = function(node) {
-  const endNode = node.nodes && privateUtil.getLast(node.nodes);
-  if (endNode && node.source && !node.source.end) {
-    node = endNode;
-  }
-
-  let loc;
-  if (node.range) {
-    loc = node.range[1];
-  } else if (typeof node.end === "number") {
-    loc = node.end;
-  }
-
-  if (node.__location) {
-    return node.__location.endOffset;
-  }
-  if (node.typeAnnotation) {
-    return Math.max(loc, locEnd(node.typeAnnotation));
-  }
-
-  if (node.loc && !loc) {
-    return node.loc.end;
-  }
-
-  return loc;
-};
 
 const languages = [
   {
@@ -179,56 +122,6 @@ const languages = [
   }
 ];
 
-const typescript = {
-  get parse() {
-    return eval("require")("./parser-typescript");
-  },
-  astFormat: "estree",
-  hasPragma,
-  locStart,
-  locEnd
-};
-
-const babylon = {
-  get parse() {
-    return eval("require")("./parser-babylon");
-  },
-  astFormat: "estree",
-  hasPragma,
-  locStart,
-  locEnd
-};
-
-const parsers = {
-  babylon,
-  json: Object.assign({}, babylon, {
-    hasPragma() {
-      return true;
-    }
-  }),
-  json5: babylon,
-  "json-stringify": {
-    get parse() {
-      return eval("require")("./parser-json-stringify");
-    },
-    astFormat: "estree-json",
-    locStart,
-    locEnd
-  },
-  flow: {
-    get parse() {
-      return eval("require")("./parser-flow");
-    },
-    astFormat: "estree",
-    hasPragma,
-    locStart,
-    locEnd
-  },
-  "typescript-eslint": typescript,
-  // TODO: Delete this in 2.0
-  typescript
-};
-
 const printers = {
   estree: estreePrinter,
   "estree-json": estreeJsonPrinter
@@ -237,6 +130,5 @@ const printers = {
 module.exports = {
   languages,
   options,
-  parsers,
   printers
 };
