+++ src/language-js/loc.js
@@ -1,0 +1,63 @@
+"use strict";
+
+const getLast = require("../utils/get-last");
+
+const locStart = function(node) {
+  // Handle nodes with decorators. They should start at the first decorator
+  if (
+    node.declaration &&
+    node.declaration.decorators &&
+    node.declaration.decorators.length > 0
+  ) {
+    return locStart(node.declaration.decorators[0]);
+  }
+  if (node.decorators && node.decorators.length > 0) {
+    return locStart(node.decorators[0]);
+  }
+
+  if (node.__location) {
+    return node.__location.startOffset;
+  }
+  if (node.range) {
+    return node.range[0];
+  }
+  if (typeof node.start === "number") {
+    return node.start;
+  }
+  if (node.loc) {
+    return node.loc.start;
+  }
+  return null;
+};
+
+const locEnd = function(node) {
+  const endNode = node.nodes && getLast(node.nodes);
+  if (endNode && node.source && !node.source.end) {
+    node = endNode;
+  }
+
+  let loc;
+  if (node.range) {
+    loc = node.range[1];
+  } else if (typeof node.end === "number") {
+    loc = node.end;
+  }
+
+  if (node.__location) {
+    return node.__location.endOffset;
+  }
+  if (node.typeAnnotation) {
+    return Math.max(loc, locEnd(node.typeAnnotation));
+  }
+
+  if (node.loc && !loc) {
+    return node.loc.end;
+  }
+
+  return loc;
+};
+
+module.exports = {
+  locStart,
+  locEnd
+};
