--- src/language-js/parser-json-stringify.js
@@ -1,77 +1,0 @@
-"use strict";
-
-const parserBabylon = eval("require")("./parser-babylon");
-const createError = require("../common/parser-create-error");
-
-function parse(text, parsers, opts) {
-  const ast = parserBabylon(
-    text,
-    parsers,
-    Object.assign({}, opts, { parser: "json" })
-  );
-
-  ast.comments.forEach(assertJsonNode);
-  assertJsonNode(ast);
-
-  return ast;
-}
-
-function assertJsonNode(node, parent) {
-  switch (node.type) {
-    case "ArrayExpression":
-      return node.elements.forEach(assertJsonChildNode);
-    case "ObjectExpression":
-      return node.properties.forEach(assertJsonChildNode);
-    case "ObjectProperty":
-      // istanbul ignore if
-      if (node.computed) {
-        throw createJsonError("computed");
-      }
-      // istanbul ignore if
-      if (node.shorthand) {
-        throw createJsonError("shorthand");
-      }
-      return [node.key, node.value].forEach(assertJsonChildNode);
-    case "UnaryExpression":
-      switch (node.operator) {
-        case "+":
-        case "-":
-          return assertJsonChildNode(node.argument);
-        // istanbul ignore next
-        default:
-          throw createJsonError("operator");
-      }
-    case "Identifier":
-      if (parent && parent.type === "ObjectProperty" && parent.key === node) {
-        return;
-      }
-      throw createJsonError();
-    case "NullLiteral":
-    case "BooleanLiteral":
-    case "NumericLiteral":
-    case "StringLiteral":
-      return;
-    // istanbul ignore next
-    default:
-      throw createJsonError();
-  }
-
-  function assertJsonChildNode(child) {
-    return assertJsonNode(child, node);
-  }
-
-  // istanbul ignore next
-  function createJsonError(attribute) {
-    const name = !attribute
-      ? node.type
-      : `${node.type} with ${attribute}=${JSON.stringify(node[attribute])}`;
-    return createError(`${name} is not allowed in JSON.`, {
-      start: {
-        line: node.loc.start.line,
-        column: node.loc.start.column + 1
-      }
-    });
-  }
-}
-
-module.exports = parse;
