+++ src/main/ast-to-doc.js
@@ -87,7 +87,12 @@ function genericPrint(path, options, printPath, args) {
   if (node) {
     try {
       // Potentially switch to a different parser
-      const sub = multiparser.printSubtree(path, printPath, options);
+      const sub = multiparser.printSubtree(
+        path,
+        printPath,
+        options,
+        printAstToDoc
+      );
       if (sub) {
         return sub;
       }
