+++ src/utils/get-last.js
@@ -1,0 +1,5 @@
+"use strict";
+
+module.exports = function(arr) {
+  return arr.length > 0 ? arr[arr.length - 1] : null;
+};
